rf: make bottom nav bar re-usable
