+++ lib/screens/intro_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_sample/shared/menu_drawer.dart';
 
+import '../shared/menu_bottom.dart';
+
 class IntroScreen extends StatelessWidget {
   const IntroScreen({Key? key}) : super(key: key);
 
@@ -9,23 +11,7 @@ class IntroScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(title: const Text('Naki Fitness')),
       drawer: const MenuDrawer(),
-      bottomNavigationBar: BottomNavigationBar(
-          onTap: (int index) {
-            switch (index) {
-              case 0:
-                Navigator.pushNamed(context, '/');
-                break;
-              case 1:
-                Navigator.pushNamed(context, '/bmi');
-                break;
-              default:
-            }
-          },
-          items: const [
-            BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
-            BottomNavigationBarItem(
-                icon: Icon(Icons.monitor_weight), label: 'BMI')
-          ]),
+      bottomNavigationBar: const MenuBottom(),
       body: Container(
           decoration: const BoxDecoration(
               image: DecorationImage(
